feat/water-2992 -- start date form bug fixed -- clean object moved to shared -- h.postRedirectGet uses clean object to remove empty params
