+++ src/shared/lib/clean-object.js
